+++ lib/app/theme/app_theme.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:aedify/app/theme/app_colors.dart';
+import 'package:aedify/shared/theme/app_colors.dart';
 import 'package:aedify/shared/theme/app_spacing.dart';
 import 'package:aedify/shared/theme/app_text_styles.dart';