+++ lib/shared/constants/svg_assets_bodymap.dart
@@ -1,0 +1,6 @@
+class BodymapSvgAssets {
+  BodymapSvgAssets._();
+
+  static const String front = 'assets/svgs/bodymap/front.svg';
+  static const String back = 'assets/svgs/bodymap/back.svg';
+}